+++ e2e/draw.spec.ts
@@ -2,7 +2,8 @@ import { test, expect } from '@playwright/test'
 import { openSnowbox } from './utils/openSnowbox'
 import { draw, drag } from './utils/draw'
 
-const drawTargetId = '#vuex-target-draw-result'
+const drawTargetId = 'vuex-target-draw-result'
+const drawTargetLocatorString = `#${drawTargetId}`
 
 test('the draw tool can be used to draw polygons on the map', async ({
   page,
@@ -12,7 +13,7 @@ test('the draw tool can be used to draw polygons on the map', async ({
 
   await page.getByLabel('Draw tools').click()
   await page.getByText('Draw, write and measure').click()
-  await page.getByText('Polygon').click()
+  await page.getByText('Polygon', { exact: true }).click()
 
   if (isMobile) {
     // menu gets in the way in mobile
@@ -29,7 +30,9 @@ test('the draw tool can be used to draw polygons on the map', async ({
     [40, -40, 'dblclick'],
   ])
 
-  const drawing = JSON.parse(await page.locator(drawTargetId).innerText())
+  const drawing = JSON.parse(
+    await page.locator(drawTargetLocatorString).innerText()
+  )
 
   expect(drawing.type).toBe('FeatureCollection')
   expect(drawing.features.length).toBe(1)
@@ -74,7 +77,9 @@ test('two features drawn at the same coordinate can be modified separately', asy
     [-40, -40],
   ])
 
-  const drawing = JSON.parse(await page.locator(drawTargetId).innerText())
+  const drawing = JSON.parse(
+    await page.locator(drawTargetLocatorString).innerText()
+  )
 
   expect(drawing.type).toBe('FeatureCollection')
   expect(drawing.features.length).toBe(2)
@@ -83,5 +88,148 @@ test('two features drawn at the same coordinate can be modified separately', asy
   )
   expect(drawing.features[0].geometry.coordinates[1]).not.toBe(
     drawing.features[1].geometry.coordinates[1]
+  )
+})
+
+test('drawn polygons can be cut', async ({ page, isMobile }) => {
+  await openSnowbox(page)
+
+  await page.getByLabel('Draw tools').click()
+  await page.getByText('Draw, write and measure').click()
+  await page.getByText('Polygon', { exact: true }).click()
+
+  if (isMobile) {
+    // menu gets in the way in mobile
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
+  }
+
+  await draw(page, [
+    [-40, 0, 'click'],
+    [40, -40, 'click'],
+    [40, 40, 'click'],
+    [-40, 40, 'dblclick'],
+  ])
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(1)
+  }
+
+  await page.getByText('Cut polygons').click()
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
+  }
+
+  await draw(page, [
+    [-80, 0, 'click'],
+    [160, 0, 'dblclick'],
+  ])
+
+  await page.waitForFunction(
+    (id) =>
+      document.getElementById(id)?.innerText !== 'uninitialized' &&
+      JSON.parse(document.getElementById(id)?.innerText || '').features
+        .length === 2,
+    drawTargetId,
+    { timeout: 5000 }
+  )
+})
+
+test('drawn polygons can be duplicated', async ({ page, isMobile }) => {
+  await openSnowbox(page)
+
+  await page.getByLabel('Draw tools').click()
+  await page.getByText('Draw, write and measure').click()
+  await page.getByText('Polygon', { exact: true }).click()
+
+  if (isMobile) {
+    // menu gets in the way in mobile
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
+  }
+
+  await draw(page, [
+    [-40, 0, 'click'],
+    [40, -40, 'click'],
+    [40, 40, 'dblclick'],
+  ])
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(1)
+  }
+
+  await page.getByText('Duplicate').click()
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
+  }
+
+  await draw(page, [[0, -20, 'click']])
+
+  await page.waitForFunction(
+    (id) =>
+      document.getElementById(id)?.innerText !== 'uninitialized' &&
+      JSON.parse(document.getElementById(id)?.innerText || '').features
+        .length === 2,
+    drawTargetId,
+    { timeout: 5000 }
+  )
+})
+
+test('drawn polygons can be merged', async ({ page, isMobile }) => {
+  await openSnowbox(page)
+
+  await page.getByLabel('Draw tools').click()
+  await page.getByText('Draw, write and measure').click()
+  await page.getByText('Polygon', { exact: true }).click()
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
+  }
+
+  await draw(page, [
+    [-40, 0, 'click'],
+    [0, -40, 'click'],
+    [-40, 40, 'dblclick'],
+  ])
+
+  await draw(page, [
+    [40, 0, 'click'],
+    [40, 0, 'click'],
+    [-40, -40, 'dblclick'],
+  ])
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(1)
+  }
+
+  await page.getByText('Merge polygons').click()
+
+  if (isMobile) {
+    await page.getByLabel('Draw tools').click()
+    await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
+  }
+
+  await draw(page, [
+    [60, -20, 'click'],
+    [-120, 0, 'click'],
+    [0, 40, 'click'],
+    [120, 0, 'dblclick'],
+  ])
+
+  await page.waitForFunction(
+    (id) =>
+      document.getElementById(id)?.innerText !== 'uninitialized' &&
+      JSON.parse(document.getElementById(id)?.innerText || '').features
+        .length === 1,
+    drawTargetId,
+    { timeout: 5000 }
   )
 })